+++ tests/testthat/test_emuR-query.database.R
@@ -39,8 +39,7 @@ test_that("Load example database ae",{
   })
   
   test_that("Query level label groups",{
-    
-    
+
     sl1=query(ae,"Phoneme=nasal",resultType='emusegs')
     # TODO check some items
     expect_that(nrow(sl1),equals(23))
@@ -434,7 +433,7 @@ test_that("Load example database ae",{
     expect_equal(nrow(sl), 10)
     sl  = query (ae ,  "[[[Phonetic = n -> Phonetic =z] -> Phonetic = S ] ^ [Text = friends -> Text = she]]")
     expect_equal(sl$labels, "n->z->S")
-    sl = query(ae, "[Utterance =~ .* ^ Phonetic == @]")
+    sl = query(ae, "[Utterance =~ .* ^ Phonetic == @]", verbose = F)
     expect_equal(nrow(sl), 7)
     expect_equal(sl$labels[1], "")
   })
@@ -458,7 +457,7 @@ test_that("Load example database ae",{
 
   test_that("correct times are calculated for Intonational",{
     skip_on_cran()
-    sl = query(ae, "Intonational == L%", timeRefSegmentLevel = "Phonetic")
+    sl = query(ae, "Intonational == L%", timeRefSegmentLevel = "Phonetic", verbose = F)
     all(round(sl$start, 3) == round(c(256.925, 571.925, 379.525, 425.375, 299.975, 513.925, 475.775), 3))
     all(round(sl$end, 3) == round(c(2604.425, 2753.975, 2692.325, 3456.825, 2469.525, 2554.175, 2794.925), 3))
   })